update auth handling to app and components to single source
